+++ geographic_services/service_area/tests.py
@@ -1,7 +1,10 @@
 import json
+from unittest import mock
 
 import pytest
+from rest_framework import status
 
+from geographic_services.service_area.models import ServiceArea
 from geographic_services.service_area.views import ServiceAreaView
 
 
@@ -16,7 +19,7 @@ def test_should_save_new_service_area_successfully(
     )
     response = view(request)
 
-    assert response.status_code == 201
+    assert response.status_code == status.HTTP_201_CREATED
     del response.data['service_area_id']
     assert response.data == service_area_payload
 
@@ -38,7 +41,7 @@ def test_post_service_area_should_return_bad_request_when_payload_invalid(
         content_type='application/json'
     )
     response = view(request)
-    assert response.status_code == 400
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
 
 
 def test_put_should_update_service_area_successfully(
@@ -55,7 +58,7 @@ def test_put_should_update_service_area_successfully(
     )
     response = view(request, service_area_id=service_area_id)
 
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     assert response.data['price'] == '20.50'
 
 
@@ -84,7 +87,7 @@ def test_put_service_area_should_return_bad_request_when_payload_is_invalid(
     )
     response = view(request, service_area_id=service_area_id)
 
-    assert response.status_code == 400
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
 
 
 def test_put_service_area_should_return_not_found(
@@ -98,7 +101,7 @@ def test_put_service_area_should_return_not_found(
     )
     response = view(request, service_area_id='invalid-id')
 
-    assert response.status_code == 404
+    assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 def test_patch_should_update_service_area_successfully(
@@ -114,7 +117,7 @@ def test_patch_should_update_service_area_successfully(
         content_type='application/json'
     )
     response = view(request, service_area_id=service_area_id)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     assert response.data['price'] == '30.00'
 
 
@@ -141,7 +144,7 @@ def test_patch_service_area_should_return_bad_request_when_payload_is_invalid(
     )
     response = view(request, service_area_id=service_area_id)
 
-    assert response.status_code == 400
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
 
 
 def test_patch_service_area_should_return_not_found(
@@ -155,7 +158,7 @@ def test_patch_service_area_should_return_not_found(
     )
     response = view(request, service_area_id='invalid-id')
 
-    assert response.status_code == 404
+    assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 def test_delete_should_remove_service_area_successfully(
@@ -169,7 +172,7 @@ def test_delete_should_remove_service_area_successfully(
         content_type='application/json'
     )
     response = view(request, service_area_id=service_area_id)
-    assert response.status_code == 204
+    assert response.status_code == status.HTTP_204_NO_CONTENT
 
 
 def test_delete_service_area_should_return_not_found(
@@ -181,7 +184,7 @@ def test_delete_service_area_should_return_not_found(
         content_type='application/json'
     )
     response = view(request, service_area_id='invalid-id')
-    assert response.status_code == 404
+    assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
 def test_list_should_retrieve_all_service_areas_successfully(
@@ -193,7 +196,7 @@ def test_list_should_retrieve_all_service_areas_successfully(
         content_type='application/json'
     )
     response = view(request)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     content = response.render().content
     content = json.loads(content)
     assert content['count'] == 15
@@ -208,7 +211,7 @@ def test_list_service_areas_should_retrieve_empty_result_successfully(
         content_type='application/json'
     )
     response = view(request)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     content = response.render().content
     content = json.loads(content)
     assert content['count'] == 0
@@ -225,7 +228,7 @@ def test_get_should_retrieve_service_area_successfully(
         content_type='application/json'
     )
     response = view(request, service_area_id=service_area_id)
-    assert response.status_code == 200
+    assert response.status_code == status.HTTP_200_OK
     content = response.render().content
     content = json.loads(content)
     assert content == saved_service_area
@@ -240,4 +243,27 @@ def test_get_service_area_should_return_not_found(
         content_type='application/json'
     )
     response = view(request, service_area_id='invalid-id')
-    assert response.status_code == 404
+    assert response.status_code == status.HTTP_404_NOT_FOUND
+
+
+@mock.patch.object(ServiceArea, 'objects')
+def test_get_service_areas_intersections_should_return_service_area(
+    mock_service, client, db_setup, saved_list_service_area
+):
+    ''''
+    For this test was necessary to mock the query database because
+    intersects is not implemented in mongomock
+    '''
+    mock_service.return_value = saved_list_service_area
+
+    view = ServiceAreaView.as_view(
+        {'get': 'get_service_areas_intersections'}, name='intersections'
+    )
+    request = client.get(
+        '/v1/service_areas/list_point_intersections?'
+        'latitude=121.00380420684814&longitude=14.515791721361213',
+        content_type='application/json'
+    )
+
+    response = view(request)
+    assert response.status_code == status.HTTP_200_OK